requirements.txt
fastapi
mangum
supabase
pinecone-client
httpx
trafilatura
langchain-text-splitters
python-dotenv
boto3
